requirements.txt
# Core dependencies
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
requests>=2.28.2
python-dotenv>=1.0.0
apscheduler>=3.10.1
fastmcp>=2.2.0

# Data analysis
pandas>=1.5.3
numpy>=1.24.3
scikit-learn>=1.2.2

# Database
sqlalchemy>=2.0.9
alembic>=1.10.3
aiosqlite>=0.19.0

# GitHub integration
PyGithub>=1.58.2

# Web integrations
httpx>=0.24.0
aiohttp>=3.8.4

# UI
jinja2>=3.1.2
itsdangerous>=2.1.2

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0

# Development
black>=23.3.0
isort>=5.12.0
pylint>=2.17.2

# Core AI Agent 框架
crewai==0.22.5
autogen-agentchat==0.2.0
semantic-kernel==0.9.0

# Web 框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 数据库和缓存
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
asyncpg==0.29.0

# 数据处理和分析
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
scipy==1.11.4

# AI 和 LLM
openai==1.6.1
anthropic==0.7.8
transformers==4.36.2
torch==2.1.1
sentence-transformers==2.2.2

# GitHub 和社交媒体 API
PyGithub==2.1.1
tweepy==4.14.0
requests==2.31.0
aiohttp==3.9.1

# 任务队列和调度
celery==5.3.4
apscheduler==3.10.4
dramatiq==1.15.0

# 监控和日志
loguru==0.7.2
prometheus-client==0.19.0
sentry-sdk==1.38.0

# 配置和工具
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1 